% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_plot.R
\name{accuracy_plot}
\alias{accuracy_plot}
\title{Creates a scatter plot}
\usage{
accuracy_plot(predictions, pred_var)
}
\arguments{
\item{predictions}{a dataframe containing predictions from a linear model fit}

\item{pred_var}{an unquoted column name}
}
\value{
A scatter plot of two numerical features
}
\description{
Creates a scatter plot from the provided data comparing the accuracy of
a regression model to the actual observations
}
\examples{
accuracy_plot(fatal_predictions, fatalities) + 
labs(x = "Actual Number of Fatalities", y = "Predicted Number of Fatalities")
}
