% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_viz.R
\name{boxplot_viz}
\alias{boxplot_viz}
\title{Boxplot Viz}
\usage{
boxplot_viz(data_frame, col)
}
\arguments{
\item{data_frame}{a data frame or data frame extension}

\item{col}{an unquoted column name from the data frame}
}
\value{
A boxplot of single variable of interest
}
\description{
Creates a boxplot visualization from a specified column.
}
\examples{
boxplot_viz(test_data, fatalities)
}
