% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scatterplot.R
\name{create_scatterplot}
\alias{create_scatterplot}
\title{Creates a scatter plot}
\usage{
create_scatterplot(data, x_col, y_col)
}
\arguments{
\item{data}{data frame or tibble}

\item{x_col}{unquoted column name to plot on the x-axis from data}

\item{y_col}{unquoted column name to plot on the y-axis from data}
}
\value{
A scatter plot of two numerical features
}
\description{
Creates a scatter plot from the provided data and specified two numerical
columns
}
\examples{
create_scatterplot(tornado_train_data, width, fatalities) + 
labs(x = "Width of tornadoes (yards)", y = "Fatalities", 
title = "Figure 2: Scatterplot of width (yards) of tornado and fatalities")
}
