% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_linear_model.R
\name{fit_linear_model}
\alias{fit_linear_model}
\title{Fit Linear Model}
\usage{
fit_linear_model(formula, data)
}
\arguments{
\item{formula}{The model formula.}

\item{data}{The training data.}
}
\value{
A fitted model workflow object.
}
\description{
Creates and fits a linear model to the training data with linear regression.
}
\examples{
fit_linear_model(fatalities ~ width + length, train_data)
}
