% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process Data}
\usage{
process_data(data_frame)
}
\arguments{
\item{data_frame}{A data frame or data frame extension}
}
\value{
A data frame with no NA values. Column names should be in this order from left to right: ID, year, month, day, time,
datetime_utc, state, mag, injuries, fatalities, start_lat, start_lon, end_lat, end_lon, length, width, ns
}
\description{
Removes NA values in \code{mag} variable and repetitive/irrelevant columns from data frame; renames column names
in this order given that original columns in dataset are in correct order: ID, year, month, day, time, datetime_utc, state, mag,
injuries, fatalities, start_lat, start_lon, end_lat, end_lon, length, width, ns
}
\examples{
process_data(raw_data)
}
